Modular closing chat, and other improvements.  Passed build.
