+++ charlie_chat/app/globals.css
@@ -116,10 +116,12 @@
   * {
     @apply border-border outline-ring/50;
   }
+
   body {
     @apply bg-background text-foreground;
   }
 }
+
 @media print {
   .print-assumptions {
     page-break-before: always;
